consumer/requirements.txt
fastapi==0.81.0
uvicorn==0.18.3
kafka-python==2.0.2
elasticsearch==8.4.0